upd
git-svn-id: https://fv3.googlecode.com/svn/trunk@54 1faddcec-d5ef-11de-96bf-fbce4fd84a6e

+++ src/fv3/ds/AmbientColorNode.java
@@ -26,7 +26,7 @@ public final class AmbientColorNode
     extends Node
 {
 
-    public float[] color = new float[3];
+    public double[] color = new double[3];
     public Track   color_track = new Track(Track.Type.VECTOR);
