+++ src/fv3/ds/OmnilightNode.java
@@ -26,8 +26,8 @@ public final class OmnilightNode
     extends Node
 {
 
-    public float[]   pos = new float[3];
-    public float[]   color = new float[3];
+    public double[]   pos = new double[3];
+    public double[]   color = new double[3];
     public Track     pos_track = new Track(Track.Type.VECTOR);
     public Track     color_track = new Track(Track.Type.VECTOR);
